- First version of Checkbox - Added 'Join forum' help menu item - Restructured help menu - about.html and css reworked to add 'powered by zest' - jtracer.properties finalized - manifest populated through pom.xml

+++ src/main/java/net/libcsdbg/jtracer/component/Alert.java
@@ -110,7 +110,7 @@ public class Alert extends JDialog implements ActionListener,
 			center.add(l);
 
 			if (i == 0) {
-				bagConstraints.insets.top = componentSvc.getLineSpacing("alert");
+				bagConstraints.insets.top = componentSvc.getLineSpacing("alert").intValue();
 			}
 		}
