+++ src/main/java/net/libcsdbg/jtracer/component/InputPrompt.java
@@ -92,7 +92,7 @@ public class InputPrompt extends JDialog implements ActionListener,
 			center.add(l);
 
 			if (i == 0) {
-				bagConstraints.insets.top = componentSvc.getLineSpacing("alert");
+				bagConstraints.insets.top = componentSvc.getLineSpacing("alert").intValue();
 			}
 		}
